feat: update workshop triggered by post to /update-workshop
