+++ backend/app/src/main/kotlin/backend/App.kt
@@ -20,6 +20,4 @@ fun Application.module() {
     ).init()
     configureAuth()
     configureRouting()
-
-    startWorkshopDatabaseUpdateTask()
 }
