+++ backend/app/src/main/kotlin/backend/config/Routing.kt
@@ -5,6 +5,7 @@ import backend.repository.SpeakerRepository
 import backend.repository.UserRepository
 import backend.repository.WorkshopRepository
 import backend.route.*
+import backend.service.WorkshopService
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
 import io.ktor.server.response.*
@@ -15,10 +16,11 @@ fun Application.configureRouting() {
     val adminRepository = AdminRepository()
     val workshopRepository = WorkshopRepository()
     val speakerRepository = SpeakerRepository()
+    val workshopService = WorkshopService(environment.config)
 
     configureAuth0Route(userRepository)
     configureUserRoutes(userRepository)
-    configureWorkshopRoutes(workshopRepository, speakerRepository)
+    configureWorkshopRoutes(workshopRepository, speakerRepository, workshopService)
     configureAdminRoutes(adminRepository)
     configureApiRoutes()
