+++ backend/app/src/main/kotlin/backend/route/WorkshopRoute.kt
@@ -1,21 +1,27 @@
 package backend.route
 
-import backend.repository.Speaker
+import backend.service.WorkshopService
 import backend.repository.SpeakerRepository
-import backend.repository.Workshop
 import backend.repository.WorkshopRepository
+import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
-fun Application.configureWorkshopRoutes(workshopRepository: WorkshopRepository, speakerRepository: SpeakerRepository) {
+fun Application.configureWorkshopRoutes(workshopRepository: WorkshopRepository,
+                                        speakerRepository: SpeakerRepository,
+                                        workshopService: WorkshopService
+) {
     routing {
-        workshopRoute(workshopRepository, speakerRepository)
+        workshopRoute(workshopRepository, speakerRepository, workshopService)
     }
 }
 
-fun Routing.workshopRoute(workshopRepository: WorkshopRepository, speakerRepository: SpeakerRepository) {
+fun Routing.workshopRoute(workshopRepository: WorkshopRepository,
+                          speakerRepository: SpeakerRepository,
+                          workshopService: WorkshopService
+) {
     authenticate("auth0-user") {
         get("/workshop") {
             val speakers = speakerRepository.list()
@@ -26,6 +32,11 @@ fun Routing.workshopRoute(workshopRepository: WorkshopRepository, speakerReposit
             call.respond(workshops)
         }
     }
-
+    authenticate("auth0-admin") {
+        post("/update-workshop") {
+            workshopService.workshopDatabaseUpdate()
+            call.respond(HttpStatusCode.OK)
+        }
+    }
 }
